move code from fwdpy11 namespace to core lib (non-public)
* delete lots of code supporting the old API
* refactor some code from "header only" to .hpp/.cc
* cleanup commented-out code and outdated comments
