+++ lib/evolve_discrete_demes/discrete_demography/simulation/pick_parents.hpp
@@ -22,15 +22,13 @@
 
 #include <cstdint>
 #include <sstream>
-#include "../../rng.hpp"
+#include <fwdpy11/rng.hpp>
 #include "core/demes/forward_graph.hpp"
 #include "fwdpp/gsl_discrete.hpp"
-#include "migration_lookup.hpp"
-#include "deme_property_types.hpp"
 #include "multideme_fitness_lookups.hpp"
 #include "mating_event_type.hpp"
 
-namespace fwdpy11
+namespace fwdpy11_core
 {
     namespace discrete_demography
     {
@@ -47,32 +45,14 @@ namespace fwdpy11
             }
         };
 
-        inline parent_data
-        pick_parents(const GSLrng_t& rng, const std::int32_t offspring_deme,
+        parent_data
+        pick_parents(const fwdpy11::GSLrng_t& rng, const std::int32_t offspring_deme,
                      const fwdpy11_core::ForwardDemesGraph& demography,
                      const fwdpp::gsl_ran_discrete_t_ptr& ancestor_deme_lookup,
                      const multideme_fitness_bookmark& fitness_bookmark,
-                     const multideme_fitness_lookups<std::uint32_t>& wlookups)
-        {
-            std::int32_t pdeme = static_cast<std::int32_t>(
-                gsl_ran_discrete(rng.get(), ancestor_deme_lookup.get()));
-
-            auto selfing_rates = demography.offspring_selfing_rates();
-            auto selfing_rate_offspring_deme
-                = *(std::begin(selfing_rates) + offspring_deme);
-
-            auto p1 = wlookups.get_parent(rng, fitness_bookmark, pdeme);
-            // FIXME: this gives rise to residual selfing
-            if (selfing_rate_offspring_deme > 0.
-                && gsl_rng_uniform(rng.get()) <= selfing_rate_offspring_deme)
-                {
-                    return {p1, p1, pdeme, pdeme, mating_event_type::selfing};
-                }
-            auto p2 = wlookups.get_parent(rng, fitness_bookmark, pdeme);
-            return {p1, p2, pdeme, pdeme, mating_event_type::outcrossing};
-        }
+                     const multideme_fitness_lookups<std::uint32_t>& wlookups);
 
     } // namespace discrete_demography
-} // namespace fwdpy11
+} // namespace fwdpy11_core
 
 #endif
